feat: stripped for cse in shop
In order for us to determine whether a card is a 'business' or a
'consumer' card, the card details need to be filled in inside our shop.
Using input fields inside Vue is not safe enough as we do not want to
save the card details in Vue logic. This is a hacky POC to show how it
could be done using an iframe. The iframe generates a encrypted string
which the PSP can decrypt.